cut api related code to new ctb_crawl.py - except code inside prepare_data - broken now
